+++ projects/packages/connection/src/class-package-version.php
@@ -12,7 +12,7 @@ namespace Automattic\Jetpack\Connection;
  */
 class Package_Version {
 
-	const PACKAGE_VERSION = '1.58.3-alpha';
+	const PACKAGE_VERSION = '1.60.1-alpha';
 
 	const PACKAGE_SLUG = 'connection';