+++ projects/plugins/jetpack/extensions/plugins/ai-assistant-plugin/components/ai-assistant-plugin-sidebar/index.tsx
@@ -7,11 +7,12 @@ import { Button, PanelBody, PanelRow, BaseControl } from '@wordpress/components'
 import { PluginPrePublishPanel } from '@wordpress/edit-post';
 import { createInterpolateElement, useCallback } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
+import React from 'react';
 /**
  * Internal dependencies
  */
 import useAICheckout from '../../../../blocks/ai-assistant/hooks/use-ai-checkout';
-import useAIFeature, {
+import useAiFeature, {
 	UpgradeTypeProp,
 } from '../../../../blocks/ai-assistant/hooks/use-ai-feature';
 import JetpackPluginSidebar from '../../../../shared/jetpack-plugin-sidebar';
@@ -65,7 +66,7 @@ const Upgrade = ( {
 };
 
 export default function AiAssistantPluginSidebar() {
-	const { requireUpgrade, upgradeType } = useAIFeature();
+	const { requireUpgrade, upgradeType } = useAiFeature();
 	const { autosaveAndRedirect, isRedirecting } = useAICheckout();
 
 	const title = __( 'AI Assistant', 'jetpack' );
@@ -94,13 +95,16 @@ export default function AiAssistantPluginSidebar() {
 					) }
 				</PanelBody>
 			</JetpackPluginSidebar>
+
 			<PluginPrePublishPanel
 				title={ title }
 				icon={ <JetpackEditorPanelLogo /> }
 				initialOpen={ false }
 			>
-				<Proofread busy={ isRedirecting } disabled={ requireUpgrade } />
-				{ requireUpgrade && <Upgrade onClick={ autosaveAndRedirect } type={ upgradeType } /> }
+				<>
+					<Proofread busy={ isRedirecting } disabled={ requireUpgrade } />
+					{ requireUpgrade && <Upgrade onClick={ autosaveAndRedirect } type={ upgradeType } /> }
+				</>
 			</PluginPrePublishPanel>
 		</>
 	);