+++ tools/e2e-commons/pages/wp-admin/dashboard.js
@@ -1,5 +1,5 @@
 import WpPage from '../wp-page.js';
-import { resolveSiteUrl } from '../../helpers/utils-helper.cjs';
+import { resolveSiteUrl } from '../../helpers/utils-helper.js';
 import PageActions from '../page-actions.js';
 
 export default class DashboardPage extends WpPage {