+++ projects/packages/my-jetpack/CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.12.1] - 2023-11-13
+### Changed
+- AI Assistant: Updated the text and image for the interstitial based on AI plan tiers. [#33981]
+- AI Assistant: Removed the ToS notice from the interstitial page. [#34076]
+
+## [3.12.0] - 2023-11-08
+### Added
+- Updated purchased state for VaultPress backup card on My Jetpack. [#33927]
+
+## [3.11.1] - 2023-11-03
+
+## [3.11.0] - 2023-10-30
+### Added
+- Add site data to unpurchased state of VaultPress Backup card to My Jetpack. [#33607]
+
 ## [3.10.0] - 2023-10-23
 ### Added
 - Add jetpack-plans dependency. It will be use to restore the reverted change on #33410. [#33706]
@@ -1079,6 +1094,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Created package
 
+[3.12.1]: https://github.com/Automattic/jetpack-my-jetpack/compare/3.12.0...3.12.1
+[3.12.0]: https://github.com/Automattic/jetpack-my-jetpack/compare/3.11.1...3.12.0
+[3.11.1]: https://github.com/Automattic/jetpack-my-jetpack/compare/3.11.0...3.11.1
+[3.11.0]: https://github.com/Automattic/jetpack-my-jetpack/compare/3.10.0...3.11.0
 [3.10.0]: https://github.com/Automattic/jetpack-my-jetpack/compare/3.9.1...3.10.0
 [3.9.1]: https://github.com/Automattic/jetpack-my-jetpack/compare/3.9.0...3.9.1
 [3.9.0]: https://github.com/Automattic/jetpack-my-jetpack/compare/3.8.2...3.9.0