+++ projects/packages/videopress/CHANGELOG.md
@@ -5,7 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.20.1] - 2023-11-13
+### Changed
+- Updated dependencies.
+
+## [0.20.0] - 2023-11-08
+### Added
+- Added support for gated content in VideoPress. [#32154]
+
+## [0.19.3] - 2023-11-03
+### Changed
+- Updated package dependencies. [#33904]
+
+## [0.19.2] - 2023-10-30
+### Fixed
+- VideoPress: Add ThemeProvider to Dropdown and Popover components. [#33825]
+
 ## [0.19.1] - 2023-10-24
+
 - Updated package dependencies.
 
 ## [0.19.0] - 2023-10-23
@@ -1161,6 +1178,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Created empty package [#24952]
 
+[0.20.1]: https://github.com/Automattic/jetpack-videopress/compare/v0.20.0...v0.20.1
+[0.20.0]: https://github.com/Automattic/jetpack-videopress/compare/v0.19.3...v0.20.0
+[0.19.3]: https://github.com/Automattic/jetpack-videopress/compare/v0.19.2...v0.19.3
+[0.19.2]: https://github.com/Automattic/jetpack-videopress/compare/v0.19.1...v0.19.2
 [0.19.1]: https://github.com/Automattic/jetpack-videopress/compare/v0.19.0...v0.19.1
 [0.19.0]: https://github.com/Automattic/jetpack-videopress/compare/v0.18.0...v0.19.0
 [0.18.0]: https://github.com/Automattic/jetpack-videopress/compare/v0.17.6...v0.18.0